+++ lib/tbot/cli/start_workload_identity_api.go
@@ -23,9 +23,9 @@ import (
 	"github.com/alecthomas/kingpin/v2"
 	"github.com/gravitational/trace"
 
-	"github.com/gravitational/teleport/lib/client"
 	"github.com/gravitational/teleport/lib/tbot/config"
 	"github.com/gravitational/teleport/lib/tbot/services/workloadidentity"
+	"github.com/gravitational/teleport/lib/utils/parse"
 )
 
 // WorkloadIdentityAPICommand implements `tbot start workload-identity-api` and
@@ -89,7 +89,7 @@ func (c *WorkloadIdentityAPICommand) ApplyConfig(cfg *config.BotConfig, l *slog.
 	case c.NameSelector != "":
 		svc.Selector.Name = c.NameSelector
 	case c.LabelSelector != "":
-		labels, err := client.ParseLabelSpec(c.LabelSelector)
+		labels, err := parse.LabelSelectorSpec(c.LabelSelector)
 		if err != nil {
 			return trace.Wrap(err, "parsing label-selector")
 		}
